requirements.txt
python-fasthtml
MonsterUI
beautifulsoup4
PyYAML

